[REFACTOR] Affected by refactor and removed unused imports
