+++ Eventy/app/src/main/java/com/example/eventy/events/model/EventTypeWithActivity.java
@@ -1,8 +1,7 @@
 package com.example.eventy.events.model;
 
-import com.example.eventy.model.solution.CategoryWithID;
+import com.example.eventy.solutions.model.CategoryWithID;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class EventTypeWithActivity {
