+++ Eventy/app/src/main/java/com/example/eventy/model/solution/Service.java
@@ -2,6 +2,7 @@ package com.example.eventy.model.solution;
 
 import com.example.eventy.model.enums.ReservationConfirmationType;
 import com.example.eventy.events.model.EventType;
+import com.example.eventy.solutions.model.Category;
 
 import java.util.ArrayList;
 import java.util.Objects;
